example-01/requirements.txt
psycopg2-binary
cx-Oracle
